refactor: improvements on router and storage files
