+++ docker-compose.yml
@@ -11,7 +11,7 @@ services:
       - POSTGRES_PASSWORD=product
       - POSTGRES_DB=product-service
     volumes:
-      - ./db-schema/:/docker-entrypoint-initdb.d/
+      - ../config/migrations/:/docker-entrypoint-initdb.d/
       - product_database_vol:/var/lib/postgresql/data
 
 volumes:
